backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
google-genai
chromadb
requests
python-multipart
python-jose[cryptography]
passlib[bcrypt]
youtube-transcript-api
beautifulsoup4
python-docx
pypdf
supabase
stripe
apscheduler
slowapi
